safkari/car_information/serializers.py
from rest_framework import serializers
from .models import Car,CarOwner


class CarSerializer(serializers.ModelSerializer):
    class Meta:
        model = Car
        fields = '__all__'

class CarOwnerSerializer(serializers.ModelSerializer):
    class Meta:
        model = CarOwner
        fields = '__all__'