safkari/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.4
Deprecated==1.2.13
Django==4.1.4
django-cors-headers==3.13.0
django-oauth-toolkit==2.2.0
djangorestframework==3.14.0
djangorestframework-oauth==1.1.0
idna==3.4
jwcrypto==1.4.2
oauthlib==3.2.2
Pillow==9.4.0
pycparser==2.21
pytz==2022.7
requests==2.28.1
sqlparse==0.4.3
urllib3==1.26.13
wrapt==1.14.1
